+++ src/qoback/closspad/state/drag_events.cljs
@@ -1,0 +1,30 @@
+(ns qoback.closspad.state.drag-events)
+
+;; selectign? -> si estamos seleccionando
+(defn- process-drop
+  [{:keys [forecast]} player selecting?]
+  (let [{:players/keys [selected]} forecast
+        selected (if (seq selected) selected [])]
+    (if selecting?
+      {:selected     (if (contains? (set selected) player)
+                       selected
+                       (conj selected player))}
+      {:selected     (filter #(not= % player) selected)})))
+
+(defn process
+  [state args]
+  (let [[action-name & args] args]
+    (case action-name
+      ;; tengo que guardar elemento y columna
+      :start (let [[element-val] args]
+               (-> state
+                   #_(assoc-in [:forecast :tmp/column] column-val)
+                   (assoc-in [:forecast :tmp/element] element-val)))
+      :end state
+      :drop (let [[dst-col] args
+                  player (-> state :forecast :tmp/element)
+                  {:keys [selected]}
+                  (process-drop state player (= dst-col :selected))]
+              (assoc-in state [:forecast :players/selected] selected))
+      (when goog.DEBUG
+        (.log js/console "Unknown UI event " action-name "with arguments" args)))))